pyproject.toml
[tool.poetry]
name = "chattr-gpt"
version = "0.1.0"
description = ""
authors = ["Ryan Schuetzler <ryan.schuetzler@byu.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
django-environ = "^0.8.1"
django-cors-headers = "^3.10.1"
requests = "^2.26.0"
django-extensions = "^3.1.5"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
openai = "^0.27.8"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.272"
black = "^21.12b0"
pytest = "^6.2.5"
pytest-django = "^4.5.2"
pre-commit = "^2.16.0"
django-debug-toolbar = "^3.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
